clean up random code

+++ m_random.h
@@ -40,14 +40,12 @@
 #include "doomtype.h"
 
 
-#include "doomtype.h"
-
 // Returns a number from 0 to 255,
 // from a lookup table.
-int32_t M_Random (void);
+int16_t M_Random (void);
 
 // As M_Random, but used only by the play simulation.
-int32_t P_Random (void);
+int16_t P_Random (void);
 
 // Fix randoms for demos.
 void M_ClearRandom (void);
